ATO-130: Delete RequestObjectService and move logic to OrchestrationAuthorizationService
Why?
- There is no need to have two classes to handle request validation.
- Common logic can be extracted out within OrchestrationAuthorizationService.
